refactor(server): split organization settings boundaries
