+++ packages/server/src/__tests__/context-reviewer-publisher.test.ts
@@ -14,8 +14,8 @@ import { createAgent } from "../services/agents/identity.js";
 import { bindAgentRuntimeSession } from "../services/agents/runtime/session.js";
 import { handleContextReviewerPrEvent as handleContextReviewerPrEventService } from "../services/context-tree/reviewer/github/pr.js";
 import { submitContextReviewOutcome } from "../services/context-tree/reviewer/github/publisher.js";
-import { putOrgSetting } from "../services/org-settings.js";
 import { upsertInstallationFromMetadata } from "../services/scm/github/app-installations.js";
+import { putOrgSetting } from "../services/settings/organization.js";
 import { createAdminContext, seedHealthyAgentRuntime, useTestApp } from "./helpers.js";
 
 const { privateKey: privateKeyPem } = generateKeyPairSync("rsa", {
