+++ packages/server/src/__tests__/context-tree-initialize.test.ts
@@ -7,10 +7,11 @@ import { members } from "../db/schema/members.js";
 import { organizationSettings } from "../db/schema/organization-settings.js";
 import { organizations } from "../db/schema/organizations.js";
 import { contextTreeRepoName } from "../services/context-tree/repo-provisioner.js";
+import { getOrgContextTreeBinding } from "../services/context-tree/settings.js";
 import { encryptValue } from "../services/crypto.js";
-import { getOrgContextTreeBinding, getOrgSetting, putOrgSetting } from "../services/org-settings.js";
 import { bindInstallationToOrg, upsertInstallationFromMetadata } from "../services/scm/github/app-installations.js";
 import { createGitlabConnection } from "../services/scm/gitlab/connections.js";
+import { getOrgSetting, putOrgSetting } from "../services/settings/organization.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestAdmin, useTestApp } from "./helpers.js";
