+++ packages/server/src/__tests__/github-app-webhook-route.test.ts
@@ -14,9 +14,9 @@ import { messages } from "../db/schema/messages.js";
 import { users } from "../db/schema/users.js";
 import { createChat } from "../services/chat/conversation.js";
 import * as eventDedupService from "../services/event-dedup.js";
-import { putOrgSetting } from "../services/org-settings.js";
 import * as githubAudienceService from "../services/scm/github/audience.js";
 import * as githubEntityStateService from "../services/scm/github/entity-state.js";
+import { putOrgSetting } from "../services/settings/organization.js";
 import { putTeamAgentAssignment } from "../services/team-agent-settings.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestAdmin, seedClient, seedHealthyAgentRuntime, useTestApp } from "./helpers.js";
