+++ packages/server/src/__tests__/ws-session-event.test.ts
@@ -13,7 +13,7 @@ import { createAgent } from "../services/agents/identity.js";
 import * as sessionEventService from "../services/chat/sessions/events.js";
 import * as contextTreeIoService from "../services/context-tree/io.js";
 import { buildLandingCampaignChatMetadata } from "../services/landing-campaigns/metadata.js";
-import { putOrgSetting } from "../services/org-settings.js";
+import { putOrgSetting } from "../services/settings/organization.js";
 import { resolveDefaultOrgId } from "../services/team/organization.js";
 import { uuidv7 } from "../uuid.js";
 import { createTestApp } from "./helpers.js";
