+++ packages/server/src/api/context-tree-snapshot.ts
@@ -7,22 +7,22 @@ import { requireUser } from "../scope/require-user.js";
 import { summarizeContextTreeUsage } from "../services/chat/sessions/events.js";
 import { buildContextTreeIoSummary } from "../services/context-tree/io.js";
 import {
+  getOrgContextReviewRuntime,
+  isOrgContextTreeBindingRuntimeCurrent,
+} from "../services/context-tree/settings.js";
+import {
   type ContextTreeBinding,
   contextTreeSnapshotWindowDays,
   getContextTreeSnapshot,
   isGithubRemoteBinding,
 } from "../services/context-tree/snapshot.js";
-import {
-  getOrgContextReviewRuntime,
-  isOrgContextTreeBindingRuntimeCurrent,
-  resolveUserPrimaryOrgId,
-} from "../services/org-settings.js";
 import { findInstallationByOrg } from "../services/scm/github/app-installations.js";
 import {
   type ContextTreeInstallationTokenResult,
   mintContextTreeInstallationToken,
   resolveContextTreeRecoveryAction,
 } from "../services/scm/github/app-token.js";
+import { resolveUserPrimaryOrgId } from "../services/team/default-membership.js";
 
 const querySchema = z
   .object({
