+++ packages/server/src/api/orgs/context-tree-snapshot.ts
@@ -6,12 +6,15 @@ import { requireOrgMembership } from "../../scope/require-org.js";
 import { summarizeContextTreeUsage } from "../../services/chat/sessions/events.js";
 import { buildContextTreeIoSummary } from "../../services/context-tree/io.js";
 import {
+  getOrgContextReviewRuntime,
+  isOrgContextTreeBindingRuntimeCurrent,
+} from "../../services/context-tree/settings.js";
+import {
   type ContextTreeBinding,
   contextTreeSnapshotWindowDays,
   getContextTreeSnapshot,
   isGithubRemoteBinding,
 } from "../../services/context-tree/snapshot.js";
-import { getOrgContextReviewRuntime, isOrgContextTreeBindingRuntimeCurrent } from "../../services/org-settings.js";
 import { findInstallationByOrg } from "../../services/scm/github/app-installations.js";
 import {
   type ContextTreeInstallationTokenResult,
