+++ packages/server/src/services/context-tree/activation.ts
@@ -8,8 +8,8 @@ import { eq } from "drizzle-orm";
 import type { Database } from "../../db/connection.js";
 import { gitlabConnections } from "../../db/schema/gitlab-connections.js";
 import type { OrgScope } from "../../scope/types.js";
-import { getOrgContextTreeSettingState } from "../org-settings.js";
 import { getOrganization } from "../team/organization.js";
+import { getOrgContextTreeSettingState } from "./settings.js";
 
 const CONTEXT_TREE_SETTINGS_URL = "/settings/context#binding";
